+++ cmd/libs/go2idl/go-to-protobuf/protobuf/cmd.go
@@ -91,9 +91,9 @@ func New() *Generator {
 			`k8s.io/api/storage/v1beta1`,
 			`k8s.io/api/storage/v1`,
 			`k8s.io/api/admissionregistration/v1alpha1`,
+			`k8s.io/api/admission/v1alpha1`,
 			`k8s.io/api/networking/v1`,
 			`k8s.io/kubernetes/federation/apis/federation/v1beta1`,
-			`k8s.io/kubernetes/pkg/apis/admission/v1alpha1`,
 			`k8s.io/metrics/pkg/apis/metrics/v1alpha1`,
 			`k8s.io/metrics/pkg/apis/custom_metrics/v1alpha1`,
 			`k8s.io/apiserver/pkg/apis/audit/v1alpha1`,
