+++ pkg/apis/admission/v1alpha1/helpers.go
@@ -17,6 +17,7 @@ limitations under the License.
 package v1alpha1
 
 import (
+	admissionv1alpha1 "k8s.io/api/admission/v1alpha1"
 	authenticationv1 "k8s.io/api/authentication/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -24,7 +25,7 @@ import (
 )
 
 // NewAdmissionReview returns an AdmissionReview for the provided admission.Attributes
-func NewAdmissionReview(attr admission.Attributes) AdmissionReview {
+func NewAdmissionReview(attr admission.Attributes) admissionv1alpha1.AdmissionReview {
 	gvk := attr.GetKind()
 	gvr := attr.GetResource()
 	aUserInfo := attr.GetUserInfo()
@@ -40,8 +41,8 @@ func NewAdmissionReview(attr admission.Attributes) AdmissionReview {
 		userInfo.Extra[key] = authenticationv1.ExtraValue(val)
 	}
 
-	return AdmissionReview{
-		Spec: AdmissionReviewSpec{
+	return admissionv1alpha1.AdmissionReview{
+		Spec: admissionv1alpha1.AdmissionReviewSpec{
 			Name:      attr.GetName(),
 			Namespace: attr.GetNamespace(),
 			Resource: metav1.GroupVersionResource{
