+++ plugin/pkg/admission/webhook/admission.go
@@ -26,6 +26,7 @@ import (
 
 	"github.com/golang/glog"
 
+	admissionv1alpha1 "k8s.io/api/admission/v1alpha1"
 	"k8s.io/api/admissionregistration/v1alpha1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -38,7 +39,7 @@ import (
 	clientset "k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 	"k8s.io/kubernetes/pkg/api"
-	admissionv1alpha1 "k8s.io/kubernetes/pkg/apis/admission/v1alpha1"
+	admissionv1alpha1helper "k8s.io/kubernetes/pkg/apis/admission/v1alpha1"
 	admissioninit "k8s.io/kubernetes/pkg/kubeapiserver/admission"
 	"k8s.io/kubernetes/pkg/kubeapiserver/admission/configuration"
 
@@ -213,7 +214,7 @@ func (a *GenericAdmissionWebhook) callHook(ctx context.Context, h *v1alpha1.Exte
 	}
 
 	// Make the webhook request
-	request := admissionv1alpha1.NewAdmissionReview(attr)
+	request := admissionv1alpha1helper.NewAdmissionReview(attr)
 	client, err := a.hookClient(h)
 	if err != nil {
 		return &ErrCallingWebhook{WebhookName: h.Name, Reason: err}
