+++ plugin/pkg/admission/webhook/admission_test.go
@@ -27,12 +27,12 @@ import (
 	"strings"
 	"testing"
 
+	"k8s.io/api/admission/v1alpha1"
 	registrationv1alpha1 "k8s.io/api/admissionregistration/v1alpha1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apiserver/pkg/admission"
 	"k8s.io/apiserver/pkg/authentication/user"
 	"k8s.io/kubernetes/pkg/api"
-	"k8s.io/kubernetes/pkg/apis/admission/v1alpha1"
 
 	_ "k8s.io/kubernetes/pkg/apis/admission/install"
 )
