+++ staging/src/k8s.io/api/admission/v1alpha1/doc.go
@@ -14,10 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// +k8s:conversion-gen=k8s.io/kubernetes/pkg/apis/admission
-// +k8s:conversion-gen-external-types=../../../../vendor/k8s.io/api/admission/v1alpha1
-// +k8s:defaulter-gen=TypeMeta
-// +k8s:defaulter-gen-input=../../../../vendor/k8s.io/api/admission/v1alpha1
+// +k8s:deepcopy-gen=package,register
+// +k8s:openapi-gen=false
 
 // +groupName=admission.k8s.io
-package v1alpha1 // import "k8s.io/kubernetes/pkg/apis/admission/v1alpha1"
+package v1alpha1 // import "k8s.io/api/admission/v1alpha1"
